Start drafting config-importer utility.
Improved integration tests as well around config import.
