+++ cmd/gmailctl-config-migrate/main.go
@@ -1,0 +1,5 @@
+// gmailctl-config-migrate is a tool to migrate gmailctl YAML configs to the latest version.
+package main
+
+func main() {
+}
