+++ integration_test.go
@@ -81,6 +81,38 @@ func TestIntegration(t *testing.T) {
 			icfgJSON, err := json.MarshalIndent(icfg, "", "  ")
 			require.Nil(t, err)
 
+			// There should be no diff now with upstream.
+			d2, err := apply.Diff(pres.GmailConfig, upres)
+			require.Nil(t, err)
+			assert.Empty(t, d2.FiltersDiff)
+			assert.Empty(t, d2.LabelsDiff)
+
+			// There should be no diff between the original and the imported config.
+			ipres, err := apply.FromConfig(icfg)
+			require.Nil(t, err)
+			d3, err := apply.Diff(ipres.GmailConfig, upres)
+			require.Nil(t, err)
+			assert.Empty(t, d3.FiltersDiff)
+			assert.Empty(t, d3.LabelsDiff)
+
+			// Import and convert to Jsonnet.
+			var buf bytes.Buffer
+			w := bufio.NewWriter(&buf)
+			err = rimport.MarshalJsonnet(icfg, w)
+			require.Nil(t, err)
+			err = w.Flush()
+			require.Nil(t, err)
+
+			// There should be no diff between the original and the converted config.
+			ijcfg, err := config.ReadJsonnet("", buf.Bytes())
+			require.Nil(t, err)
+			ijpres, err := apply.FromConfig(ijcfg)
+			require.Nil(t, err)
+			d4, err := apply.Diff(ijpres.GmailConfig, upres)
+			require.Nil(t, err)
+			assert.Empty(t, d4.FiltersDiff)
+			assert.Empty(t, d4.LabelsDiff)
+
 			// Compare the results with the golden files (or update the golden files).
 			if *update {
 				// Import.
